Replace LoadBalancer with ClusterIP
By default, Selenium chart will attempt to create a LoadBalancer. There
is no need for that, neither do we have a desire to expose Selenium to
the internet. What we like to do, is have a private networking in place,
and potential for us to expose the endpoint with either in-cluster
Ingress or VirtualService.