Make the stubs part of the main pay chart
Having the stubs as a subchart was nice for encapsulation but made it
difficult to reason about network policies and hosts.

I've also made sure the endtoend tests use the public URLs for services
that require it to avoid giving the Selenium Chrome nodes access to
everything in the environment (they shouldn't need it).